Merged Ice's changes to user.js and min.user.js
The update for those was supposed to happen last night but they didn't
sync for some reason :/
Anyway, we have a problem in that github doesn't supply the correct
content headers for the js files if you try to access the raw file.
My "solution" for this is just to pack the code into the greasemonkey
script, which would mean we would have to wait for greasemonkey to
automatically update those scripts once changes are put out.
However we could look into things like rawgit.com which exists to put
the correct content headers on github files.
What do you think?